+++ app/Resources/views/AdminAccount/adminGestionArticle.html.twig
@@ -3,43 +3,46 @@
 {% block adminContent %}
 
     <div class="tab-content col-md-10 col-md-offset-1">
-        {% if 1 %} <!-- if articles -->
+        {% if articles %}
             <div class="table-responsive">
                 <table class="table table-hover table-condensed">
                     <thead>
                     <tr>
                         <th>Titre</th>
                         <th>Contenu</th>
-                        <th></th>  <!-- Actions column -->
+                        <th>Date</th>
+                        <th>Action</th>
                     </tr>
                     </thead>
-                    <!-- for article in articles-->
-                    <tr>
-                        <td><a class="articleTitle" href="#">titre article</a></td>
-                        <td>contenu</td>
-                        <td>
-                            <a href="#" class="btn btn-info btn-xs" title="Edit"><span class="glyphicon glyphicon-edit"></span></a>
-                            <button type="button" class="btn btn-danger btn-xs" title="Delete" data-toggle="modal" data-target="#articleDialog"><span class="glyphicon glyphicon-remove"></span>
-                            </button>
-                            <div class="modal fade" id="" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
-                                <div class="modal-dialog">
-                                    <div class="modal-content">
-                                        <div class="modal-header">
-                                            <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
-                                            <h4 class="modal-title" id="myModalLabel">Confirmation needed</h4>
-                                        </div>
-                                        <div class="modal-body">
-                                            Do you really want to delete this article ?
-                                        </div>
-                                        <div class="modal-footer">
-                                            <button type="button" class="btn btn-default" data-dismiss="modal">Cancel</button>
-                                            <a href="#" class="btn btn-danger">Confirm</a>
-                                        </div>
-                                    </div><!-- /.modal-content -->
-                                </div><!-- /.modal-dialog -->
-                            </div><!-- /.modal -->
-                        </td>
-                    </tr>
+                    {% for article in articles %}
+                        <tr>
+                            <td class="col-sm-3"><a class="articleTitle" href="{{ path('article', {'id': article.id}) }}">{{ article.title }}</a></td>
+                            <td class="col-sm-7">{{ article.text | truncate(250) }}</td>
+                            <td class="col-sm-1">{{ article.date | date("d/m/Y") }}</td>
+                            <td class="col-sm-1">
+                                <a href="#" class="btn btn-info btn-xs" title="Edit"><span class="glyphicon glyphicon-edit"></span></a>
+                                <button type="button" class="btn btn-danger btn-xs" title="Delete" data-toggle="modal" data-target="#articleDialog{{ article.id }}"><span class="glyphicon glyphicon-remove"></span>
+                                </button>
+                                <div class="modal fade" id="articleDialog{{ article.id }}" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
+                                    <div class="modal-dialog">
+                                        <div class="modal-content">
+                                            <div class="modal-header">
+                                                <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
+                                                <h4 class="modal-title" id="myModalLabel">Confirmation requise</h4>
+                                            </div>
+                                            <div class="modal-body">
+                                                Voulez vous vraiment supprimer cette article ?
+                                            </div>
+                                            <div class="modal-footer">
+                                                <button type="button" class="btn btn-default" data-dismiss="modal">Annuler</button>
+                                                <a href="{{ path('article_remove', {'id': article.id}) }}" class="btn btn-danger">Confirmer</a>
+                                            </div>
+                                        </div><!-- /.modal-content -->
+                                    </div><!-- /.modal-dialog -->
+                                </div><!-- /.modal -->
+                            </td>
+                        </tr>
+                    {% endfor %}
                 </table>
             </div>
         {% else %}
