+++ src/AppBundle/Controller/BlogController.php
@@ -51,50 +51,4 @@ class BlogController extends Controller
             'nbPages' => $nbPages,
             'page' => $page,));
     }
-
-    /**
-     * @Route("/article/{id}", name="article")
-     */
-    public function articleAction(Article $article) //mettre param converter Article article pour supprimer les deux premières lignes
-    {
-        //On récupère tous les articles qui nous serviront pour l'affichage des articles récents
-        $articlesTotal = $this->getDoctrine()->getManager()->getRepository('AppBundle:Article')->getArticlesRecents();
-
-        return $this->render('Blog/article.html.twig', array(
-            'article' => $article,
-            'articlesTotal' =>$articlesTotal));
-    }
-
-    /**
-     * @Route("/ecriture_article", name="ecriture_article")
-     */
-    public function ecriture_articleAction(Request $request)
-    {
-        $article = new Article();
-        $form = $this->get('form.factory')->create(ArticleType::class, $article);
-
-        if ($request->isMethod('POST') && $form->handleRequest($request)->isValid())
-        {
-
-            $this->handleImage($article);
-            $article->setDate(new \DateTime);
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($article);
-            $em->flush();
-            return $this->redirectToRoute('blog');
-        }
-        return $this->render('Blog/ecriture_article.html.twig', array('form' => $form->createView()));
-    }
-
-    private function handleImage(Article $article)
-    {
-        $image   = new Picture();
-        $recuperationImage = $article->getPicture();
-        $file = $recuperationImage->getUrl();
-        $filename = md5(uniqid()).'.'.$file->guessExtension();
-        $file->move($this->getParameter('pictures_directory'),$filename);
-        $image->setUrl($filename);
-        $image->setAlt($article->getTitle());
-        $article->setPicture($image);
-    }
 }
